1.支持异步登录；2.新增表分片哈希值查询；3.新增贷款余额检查；4.使用flyway管理脚本；5.接口文档生成功能优化

+++ src/main/java/com/ssy/api/utils/system/JDBCHelper.java
@@ -165,9 +165,11 @@ public class JDBCHelper {
 		List<DBProcess> processList = processMapper.showProcessList();
 		for(DBProcess dbProcess : processList){
 			try{
-				processMapper.kill(dbProcess.getId());
-				log.info("Kill the process[{}] of database [{}]", dbProcess.getId(), DBContextHolder.getCurrentDataSource());
-				size++;
+				if(CommUtil.equals("sleep", dbProcess.getCommand().toLowerCase())){
+					processMapper.kill(dbProcess.getId());
+					log.info("Kill the process[{}] of database [{}]", dbProcess.getId(), DBContextHolder.getCurrentDataSource());
+					size++;
+				}
 			}catch (Exception e){
 				log.info(e.getMessage());
 			}
